needed to convert my ints to doubles

+++ src/main/java/com/cascadebot/cascadebot/utils/StringUtils.java
@@ -52,7 +52,7 @@ public class StringUtils {
      * @return
      */
     public static String getProgressBar(float percentage, int length, boolean percentAfter) {
-        int bars = (NumberUtils.round(percentage) / 100) * length;
+        int bars = NumberUtils.round((percentage / (double)100) * (double)length);
         return "[" + org.apache.commons.lang3.StringUtils.repeat("▬", bars) +
                 "](https://github.com/CascadeBot)" +
                 org.apache.commons.lang3.StringUtils.repeat("▬", length - bars) +
